Make chart ops insensitive to (missing) slashes
Before this change it was possible that a user provided correct
credentials for a repository, but the operator was unable to find them
due to too strict comparisons, also, the URL the operator constructed
to download a tarball of a chart was missing a path element when a
given repo URL did not end with a trailing slash.

To get around the first issue the repo URL from the `repositories.yaml`
and the repo URL from the given HelmRelease are both stripped from their
trailing slashes during comparison.

To get around the second issue a CleanRepoURL function was added and
implemented, which returns the RepoURL but ensures it always ends with
a trailing slash.